配送单统计列表增加删除编辑，增加商品上下架

+++ src/main/java/com/wangyao/company/delivery/model/Product.java
@@ -48,7 +48,7 @@ public class Product {
     /**
     * 状态:0-上架,1-下架
     */
-    private Byte states;
+    private Integer states;
 
     /**
     * 商品类型 0-斤 1-袋 2-瓶 3-桶 4-包
